Refactor out decimal APIs into a subpackage (#516)
* Update issue templates

* Delete custom.md

* refactor out decimal apis into a subpackage
