+++ internal/decimal/decimal_test.go
@@ -1,4 +1,4 @@
-package mssql
+package decimal
 
 import (
 	"math"
